requirements.txt
numpy==1.23.3
pandas==1.5.0
pyyaml
sklearn